Split oversized dispatchers into request-family helpers with characterization tests
The monolithic handle() in coordination.rs and execute() in executor.rs
were the two largest remaining Tiger Style function_length hotspots in
the handler layer. Both exceeded 300 lines with no assertions and routed
dozens of RPC variants through a single match arm, making it hard to
reason about each request family in isolation.

coordination.rs now classifies requests into Counter, SignedCounter,
Sequence, Lock, and RateLimiter families, each dispatched to a dedicated
handler with shared response builders. executor.rs splits into 14
request-group helpers (Repo, Blob, Tree, Commit, Ref, Issue, Patch,
Discussion, Fork, Mirror, Federation, GitBridge, Nostr, Diff) so the
top-level execute() is a pure routing table under the 70-line limit.

The auth capability matching in capability.rs gets the same treatment:
authorizes() and contains() are broken into per-family helpers so each
permission decision is testable independently.

usize leakage is reduced in three verified/public hotspots:
ci-core pipeline.rs, forge ref_validation.rs, and cluster discovery
helpers.rs now use u32 for counts and lengths at their API boundaries.

+++ crates/aspen-cluster/src/gossip/discovery/lifecycle.rs
@@ -532,10 +532,7 @@ impl GossipPeerDiscovery {
                         }
 
                         // Calculate backoff duration using existing pure function
-                        let backoff = calculate_backoff_duration(
-                            (consecutive_errors as usize).saturating_sub(1),
-                            &backoff_durations
-                        );
+                        let backoff = calculate_backoff_duration(consecutive_errors.saturating_sub(1), &backoff_durations);
 
                         tracing::warn!(
                             "gossip receiver error (retry {}/{}), backing off for {:?}: {}",
